feat: features sections
