+++ app/layout/scenarios/edit/features/add/add-btn/uploader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
