+++ app/layout/scenarios/edit/features/add/component.tsx
@@ -1,265 +1,258 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 
-import { Form as FormRFF, Field as FieldRFF } from 'react-final-form';
-import { useQueryClient } from 'react-query';
-import { useDispatch } from 'react-redux';
+import { useSelector, useDispatch } from 'react-redux';
 
 import { useRouter } from 'next/router';
 
 import { getScenarioEditSlice } from 'store/slices/scenarios/edit';
 
-import { ScenarioSidebarTabs, ScenarioSidebarSubTabs } from 'utils/tabs';
+import { motion } from 'framer-motion';
+import { ScenarioSidebarSubTabs, ScenarioSidebarTabs } from 'utils/tabs';
 import { mergeScenarioStatusMetaData } from 'utils/utils-scenarios';
 
-import {
-  useAllFeatures, useSaveSelectedFeatures, useSelectedFeatures,
-} from 'hooks/features';
-import { useSaveScenario, useScenario } from 'hooks/scenarios';
+import { useSelectedFeatures } from 'hooks/features';
+import { useScenario, useSaveScenario } from 'hooks/scenarios';
 
-import List from 'layout/scenarios/edit/features/add/list';
-import Toolbar from 'layout/scenarios/edit/features/add/toolbar';
+import HelpBeacon from 'layout/help/beacon';
+import Pill from 'layout/pill';
+import AddFeaturesModal from 'layout/scenarios/edit/features/add/add-btn';
+import ListFeatures from 'layout/scenarios/edit/features/list';
+import TargetFeatures from 'layout/scenarios/edit/features/targets';
 
-import Button from 'components/button';
 import Icon from 'components/icon';
-import Loading from 'components/loading';
-import Modal from 'components/modal';
+import InfoButton from 'components/info-button';
+import Steps from 'components/steps';
 
-import PLUS_SVG from 'svgs/ui/plus.svg?sprite';
+import FEATURE_ABUND_IMG from 'images/info-buttons/img_abundance_data.png';
+import FEATURE_SOCIAL_IMG from 'images/info-buttons/img_social_uses.png';
+import FEATURE_SPECIES_IMG from 'images/info-buttons/img_species_range.png';
 
-import Uploader from './uploader';
+import FEATURES_SVG from 'svgs/ui/features.svg?sprite';
 
-export interface ScenariosFeaturesAddProps { }
+export interface ScenariosSidebarEditFeaturesProps {
 
-export const ScenariosFeaturesAdd: React.FC<ScenariosFeaturesAddProps> = () => {
-  const [open, setOpen] = useState(false);
-  const [submitting, setSubmitting] = useState(false);
-  const [search, setSearch] = useState(null);
-  const [filters, setFilters] = useState({});
-  const [sort, setSort] = useState(null);
+}
 
-  const dispatch = useDispatch();
+export const ScenariosSidebarEditFeatures: React.FC<ScenariosSidebarEditFeaturesProps> = () => {
+  const [step, setStep] = useState(0);
   const { query } = useRouter();
-  const { pid, sid } = query;
+  const { sid } = query;
 
-  const queryClient = useQueryClient();
   const scenarioSlice = getScenarioEditSlice(sid);
+  const { setSubTab } = scenarioSlice.actions;
 
-  const { setFeatures } = scenarioSlice.actions;
+  const { tab } = useSelector((state) => state[`/scenarios/${sid}/edit`]);
+  const dispatch = useDispatch();
 
-  const selectedFeaturesMutation = useSaveSelectedFeatures({});
+  const { data: scenarioData } = useScenario(sid);
+  const { metadata } = scenarioData || {};
+  const { scenarioEditingMetadata } = metadata || {};
+  const {
+    subtab: metaSubtab,
+  } = scenarioEditingMetadata || {};
+
+  const {
+    data: selectedFeaturesData,
+  } = useSelectedFeatures(sid, {});
+
   const saveScenarioMutation = useSaveScenario({
     requestConfig: {
       method: 'PATCH',
     },
   });
 
-  const { data: scenarioData } = useScenario(sid);
-  const { metadata } = scenarioData || {};
-
-  const {
-    data: initialSelectedFeatures,
-  } = useSelectedFeatures(sid, {});
-
-  const {
-    data: allFeaturesData,
-    isFetched: allFeaturesIsFetched,
-  } = useAllFeatures(pid, {
-    search,
-    filters,
-    sort,
-  });
-
-  const INITIAL_VALUES = useMemo(() => {
-    if (initialSelectedFeatures) {
-      return {
-        selected: initialSelectedFeatures.map((s) => s.id),
-      };
-    }
-
-    return [];
-  }, [initialSelectedFeatures]);
-
-  const onToggleSelected = useCallback((id, input) => {
-    const { value, onChange } = input;
-    const selected = [...value];
-
-    const selectedIndex = selected.findIndex((f) => f === id);
-
-    if (selectedIndex !== -1) {
-      selected.splice(selectedIndex, 1);
-    } else {
-      selected.push(id);
-    }
-
-    onChange(selected);
-  }, []);
-
-  const onSearch = useCallback((s) => {
-    setSearch(s);
-  }, []);
-
-  const onFilters = useCallback((f) => {
-    setFilters(f);
-  }, []);
-
-  const onSort = useCallback((s) => {
-    setSort(s);
-  }, []);
-
-  const onSubmit = useCallback(({ selected }) => {
-    const selectedFeaturesData = allFeaturesData.filter((sf) => selected.includes(sf.id));
-
-    setSubmitting(true);
-    dispatch(setFeatures(selectedFeaturesData));
-
-    // Save current features then dismiss the modal
-    selectedFeaturesMutation.mutate({
+  const saveScenarioStatusOnBack = useCallback(async () => {
+    saveScenarioMutation.mutate({
       id: `${sid}`,
       data: {
-        status: 'draft',
-        features: selectedFeaturesData.map((feature) => {
-          const { marxanSettings, geoprocessingOperations } = feature;
-
-          return {
-            featureId: feature.id,
-            kind: geoprocessingOperations ? 'withGeoprocessing' : 'plain',
-            marxanSettings: marxanSettings || {
-              fpf: 1,
-              prop: 0.5,
-            },
-            ...!!geoprocessingOperations && { geoprocessingOperations },
-          };
+        metadata: mergeScenarioStatusMetaData(metadata, {
+          tab: ScenarioSidebarTabs.FEATURES,
+          subtab: ScenarioSidebarSubTabs.FEATURES_ADD,
         }),
       },
-    }, {
-      onSuccess: () => {
-        saveScenarioMutation.mutate({
-          id: `${sid}`,
-          data: {
-            metadata: mergeScenarioStatusMetaData(metadata, {
-              tab: ScenarioSidebarTabs.FEATURES,
-              subtab: ScenarioSidebarSubTabs.FEATURES_ADD,
-            }),
-          },
-        }, {
-          onSuccess: () => {
-            setSubmitting(false);
-            setOpen(false);
-          },
-          onError: () => {
-            setSubmitting(false);
-          },
-        });
-      },
-      onError: () => {
-        setSubmitting(false);
-      },
     });
-  }, [sid,
-    metadata,
-    allFeaturesData,
-    selectedFeaturesMutation,
-    saveScenarioMutation,
-    setFeatures,
-    dispatch,
-  ]);
+  }, [saveScenarioMutation, sid, metadata]);
+
+  useEffect(() => {
+    setStep(metaSubtab === ScenarioSidebarSubTabs.FEATURES_TARGET ? 1 : 0);
+  }, [metaSubtab]);
+
+  if (!scenarioData || tab !== ScenarioSidebarTabs.FEATURES) return null;
 
   return (
-    <>
-      <Button
-        theme="primary"
-        size="base"
-        onClick={() => setOpen(true)}
-      >
-        <span className="mr-3">Add features</span>
-        <Icon icon={PLUS_SVG} className="w-4 h-4" />
-      </Button>
+    <div className="flex flex-col flex-grow w-full h-full overflow-hidden">
+      <HelpBeacon
+        id="scenarios-features"
+        title="Features"
+        subtitle="Manage features"
+        content={(
+          <div className="space-y-2">
+            <p>
+              Features are everything you want to include in
+              your conservation or land/sea use plan such as
+              species ranges, habitats or ecoregions.
+            </p>
+            <p>
+              The first step requires adding features.
+              There are public features available for use that
+              you can access directly
+              or you can upload your private features as a shapefile.
+            </p>
+            <p>
+              The second step of this section requires setting a
+              conservation target and feature penalty factor for
+              each feature.
+            </p>
+            <p>
+              <i>
+                Note on privacy: The features you upload will only
+                be accessible inside your project to you and your
+                contributors. They will not be shared with other users.
+              </i>
+            </p>
 
-      <Modal
-        id="all-feaures"
-        title="All features"
-        open={open}
-        size="narrow"
-        onDismiss={() => {
-          setOpen(false);
-          queryClient.removeQueries(['all-features', pid]);
-        }}
+          </div>
+        )}
+        modifiers={['flip']}
+        tooltipPlacement="left"
       >
-        <FormRFF
-          key="features-list"
-          onSubmit={onSubmit}
-          initialValues={INITIAL_VALUES}
+        <motion.div
+          key="features"
+          className="flex flex-col min-h-0 overflow-hidden"
+          initial={{ opacity: 0, y: -10 }}
+          animate={{ opacity: 1, y: 0 }}
         >
-          {({ handleSubmit, values }) => (
-            <form onSubmit={handleSubmit} autoComplete="off" className="flex flex-col flex-grow overflow-hidden text-black">
+          <Pill selected>
+            <header className="flex items-start justify-between flex-shrink-0">
+              <div>
+                <div className="flex items-baseline space-x-4">
+                  <h2 className="text-lg font-medium font-heading">Features</h2>
+                  <Steps step={step + 1} length={2} />
+                  <InfoButton>
+                    <div>
+                      <h4 className="font-heading text-lg mb-2.5">What are features?</h4>
+                      <div className="space-y-2">
+                        <p>
+                          Features are the important habitats, species, processes,
+                          activities, and discrete areas that you want to consider
+                          in your planning process. Common feature data formats are
+                          range maps, polygons, abundances, and continuous scale or
+                          probability of occurrence maps (e.g. 0-1). Features can
+                          include more than just ecological data but also be cultural
+                          and socio-economic areas like community fishing grounds
+                          or traditional-use areas, and other human activities and
+                          industries. Every feature must have a minimum target
+                          amount set.
+                          Some examples include:
+                        </p>
+                        <img src={FEATURE_SPECIES_IMG} alt="Feature-Range" />
+                        <img src={FEATURE_ABUND_IMG} alt="Feature-Abundance" />
+                        <img src={FEATURE_SOCIAL_IMG} alt="Feature-Social" />
+                      </div>
+                    </div>
+                  </InfoButton>
+                </div>
 
-              <h2 className="flex-shrink-0 pl-8 mb-5 text-lg pr-28 font-heading">Add features to your planning area</h2>
+                <div className="flex items-center mt-2 space-x-2">
 
-              <Loading
-                visible={submitting}
-                className="absolute top-0 bottom-0 left-0 right-0 z-40 flex items-center justify-center w-full h-full bg-white bg-opacity-90"
-                iconClassName="w-10 h-10 text-primary-500"
-              />
+                  {step === 0 && (
+                    <>
+                      <Icon icon={FEATURES_SVG} className="w-4 h-4 text-gray-400" />
+                      <div className="text-xs uppercase font-heading">
+                        Features added:
+                        {' '}
+                        {selectedFeaturesData && <span className="ml-1 text-gray-400">{selectedFeaturesData.length}</span>}
+                      </div>
+                    </>
+                  )}
 
-              {/* Field to upload */}
-              <div className="mx-8 mt-3 mb-5">
-                <Uploader />
+                  {step === 1 && (
+                    <div className="text-xs uppercase font-heading">
+                      Set the target and FPF for your features.
+
+                    </div>
+
+                  )}
+                  {step === 1 && (
+                    <>
+                      <InfoButton>
+                        <div>
+                          <h4 className="font-heading text-lg mb-2.5">What is a target?</h4>
+                          <div className="space-y-2">
+                            <p>
+                              This value represents how much you want to conserve of a particular
+                              feature. In an ideal conservation, land or sea use plan,
+                              all your features meet their targets.
+                            </p>
+                            <p>
+                              You can set a default
+                              value for all of your features
+                              or you can set individual the targets separately for each feature.
+                              You can set your targets to 100% if you want the whole extent of
+                              your feature to be included in the solution.
+                            </p>
+                          </div>
+                          <h4 className="font-heading text-lg mt-5 mb-2.5">What is the FPF?</h4>
+                          <div className="space-y-2">
+                            <p>
+                              FPF stands for
+                              {' '}
+                              <b>Feature Penalty Factor</b>
+                              .
+                              A higher FPF value forces the Marxan algorithm
+                              to choose the planning units where this feature
+                              is present by applying a penalty if the target
+                              is missed, thereby increasing
+                              the cost of the solution. It comes into play when
+                              some of your targets fail to be met.
+                            </p>
+                            <p>
+                              In a typical
+                              workflow you start out with all FPF values set at
+                              1 and after checking the results, increase the FPF
+                              values for the particular features where targets have
+                              been missed.
+                            </p>
+                          </div>
+                        </div>
+                      </InfoButton>
+                    </>
+                  )}
+
+                </div>
               </div>
 
-              <Toolbar
-                search={search}
-                filters={filters}
-                sort={sort}
-                onSearch={onSearch}
-                onFilters={onFilters}
-                onSort={onSort}
-              />
-
-              <FieldRFF
-                name="selected"
-              >
-                {({ input }) => (
-                  <List
-                    search={search}
-                    filters={filters}
-                    sort={sort}
-                    selected={values.selected}
-                    onToggleSelected={(id) => {
-                      onToggleSelected(id, input);
-                    }}
-                  />
-                )}
-              </FieldRFF>
-
-              {allFeaturesIsFetched && (
-                <div className="flex justify-center flex-shrink-0 px-8 space-x-3">
-                  <Button
-                    className="w-full"
-                    theme="secondary"
-                    size="lg"
-                    onClick={() => setOpen(false)}
-
-                  >
-                    Cancel
-                  </Button>
-
-                  <Button
-                    type="submit"
-                    className="w-full"
-                    theme="primary"
-                    size="lg"
-                    disabled={submitting}
-                  >
-                    Save
-                  </Button>
-                </div>
+              {step === 0 && (
+                <AddFeaturesModal />
               )}
-            </form>
-          )}
-        </FormRFF>
-      </Modal>
-    </>
+            </header>
+
+            {step === 0 && (
+              <ListFeatures
+                onSuccess={() => {
+                  setStep(step + 1);
+                  dispatch(setSubTab(ScenarioSidebarSubTabs.FEATURES_TARGET));
+                }}
+              />
+            )}
+
+            {step === 1 && (
+              <TargetFeatures
+                onBack={() => {
+                  setStep(step - 1);
+                  dispatch(setSubTab(ScenarioSidebarSubTabs.FEATURES_ADD));
+                  saveScenarioStatusOnBack();
+                }}
+                onSuccess={() => {
+                  console.info('success');
+                }}
+              />
+            )}
+          </Pill>
+        </motion.div>
+      </HelpBeacon>
+    </div>
   );
 };
 
-export default ScenariosFeaturesAdd;
+export default ScenariosSidebarEditFeatures;
